Fix writes/reads >=64KB

+++ utility/SdFat.h
@@ -238,7 +238,7 @@ class SdFile : public Print {
     uint8_t b;
     return read(&b, 1) == 1 ? b : -1;
   }
-  int16_t read(void* buf, uint16_t nbyte);
+  int32_t read(void* buf, size_t nbyte);
   int8_t readDir(dir_t* dir);
   static uint8_t remove(SdFile* dirFile, const char* fileName);
   uint8_t remove(void);
@@ -284,7 +284,7 @@ class SdFile : public Print {
   /** \return SdVolume that contains this file. */
   SdVolume* volume(void) const {return vol_;}
   size_t write(uint8_t b);
-  size_t write(const void* buf, uint16_t nbyte);
+  size_t write(const void* buf, size_t nbyte);
   size_t write(const char* str);
   void write_P(PGM_P str);
   void writeln_P(PGM_P str);
